src/backend/pyproject.toml
[project]
name = "morpheus-backend"
version = "1.0.0"
description = "Morpheus groundwater modeling backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "Authlib==1.3.0",
  "Dynaconf==3.2.4",
  "Flask-Cors==4.0.0",
  "Flask==3.0.1",
  "SQLAlchemy==2.0.25",
  "celery==5.3.6",
  "flopy==3.7.0",
  "geopandas==1.1.2",
  "geotiff==0.2.10",
  "gunicorn==21.2.0",
  "inflection==0.5.1",
  "numcodecs==0.12.1",
  "python-magic==0.4.24",
  "openapi-core==0.19.0",
  "pandas==2.2.0",
  "pillow==10.2.0",
  "psycopg2-binary==2.9.9",
  "pyarrow==15.0.0",
  "pygeoif==1.2.0",
  "pymongo==4.6.1",
  "pyproj==3.6.1",
  "python-keycloak==3.7.0",
  "sentry-sdk[flask]==2.13.0",
  "scipy==1.12.0",
  "shapely==2.1.2",
  "simplejson==3.19.2",
  "tifffile>=2023.0.0,<2024.8.0",
  "zarr==2.12.0"
]

[project.optional-dependencies]
dev = [
  "basedpyright",
  "autoflake==2.3.1",
  "autopep8==2.3.2",
  "dep-check==3.1.0",
  "flake8==7.3.0",
  "folium==0.20.0",
  "jupyter==1.0.0",
  "jupyterlab==4.0.11",
  "matplotlib==3.10.8",
  "papermill==2.6.0",
  "pep8-naming==0.13.3",
  "pip==26.0",
  "pytest==8.4.2",
  "ruff"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/morpheus"]

[tool.uv]
# Ensure uv installs MODFLOW binaries
# Note: get-modflow command still needed after install

[tool.uv.sources]
# If you need to use local packages or git repos, configure them here

[tool.basedpyright]
include = ["src/morpheus"]
exclude = [
  "**/__pycache__",
  "**/.pytest_cache",
  "**/notebooks/*",
]
pythonVersion = "3.12"
typeCheckingMode = "basic"
ignore = []
reportAny = false
reportMissingImports = "error"
reportMissingTypeStubs = false
pythonPlatform = "Linux"
executionEnvironments = [
  { root = "src" }
]

[tool.ruff]
line-length = 180
include = ["src/morpheus/**/*.py", "src/tests/**/*.py", "src/*.py"]

[tool.ruff.format]
indent-style = "space"
quote-style = "single"
docstring-code-format = true

[tool.ruff.lint]
select = [
  # pycodestyle
  "E",
  # Pyflakes
  "F",
  # pyupgrade
  "UP",
  # flake8-bugbear
  "B",
  # flake8-simplify
  "SIM",
  # isort
  "I",
]

# Disable rules that overlap/conflict with formatting:
ignore = [
  # indentation (formatter owns this)
  "W191", # tab-indentation
  "E111", # indentation-with-invalid-multiple
  "E114", # indentation-with-invalid-multiple-comment
  "E117", # over-indented

  # docstring/quote style (formatter owns this)
  "D206", # docstring-tab-indentation
  "D300", # triple-single-quotes
  "Q000", # bad-quotes-inline-string
  "Q001", # bad-quotes-multiline-string
  "Q002", # bad-quotes-docstring
  "Q003", # avoidable-escaped-quote

  # trailing-comma rules (formatter owns this)
  "COM812", # missing-trailing-comma
  "COM819", # prohibited-trailing-comma
]
